Fix critical database alignment issues in schemas
CRITICAL DATABASE ALIGNMENT FIXES:
- Fix admin_audit schema: action is NOT NULL, target_type/target_id are nullable
- Fix credit_transactions schema: transaction_type is text (not enum), all NOT NULL fields corrected
- Fix user_roles/role_permissions: use bigint for id fields (not number)
- Fix org_settings schema: updated_at is NOT NULL in database
- Remove incorrect default values and enums that don't match database

All schemas now precisely match actual Supabase database constraints.

+++ apps/id-gen/src/schemas/auth.schema.ts
@@ -62,16 +62,16 @@ export const userProfileUpdateSchema = z.object({
 	remove_watermarks: z.boolean().optional()
 });
 
-// User roles assignment schemas (user_roles table)
+// User roles assignment schemas (user_roles table - matches database)
 export const userRoleAssignmentSchema = z.object({
-	id: z.number().optional(),
+	id: z.number().int().optional(), // bigint in database
 	user_id: z.string().uuid(),
 	role: userRoleSchema
 });
 
-// Role permissions schemas (role_permissions table)  
+// Role permissions schemas (role_permissions table - matches database)  
 export const rolePermissionSchema = z.object({
-	id: z.number().optional(),
+	id: z.number().int().optional(), // bigint in database
 	role: userRoleSchema,
 	permission: z.enum([
 		// Template permissions
